requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.9
numpy==1.24.3
pygame==2.5.2
pillow==10.1.0